+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/) and this project adheres to [Semantic Versioning](https://semver.org/)
 
+## [unreleased]
+
+### Changed
+
+- output graph label derived from input graph
+
 ## [1.0.0rc1] 2024-08-28
 
 ### Added